pyproject.toml
[tool.poetry]
name = "bloxlink-http"
version = "0.1.0"
description = "An http interactions worker for Bloxlink"
authors = ["Bloxlink <oss@blox.link>"]
license = "AGPL-3.0-only"
readme = "README.md"
homepage = "https://blox.link/"
repository = "https://github.com/bloxlink/bloxlink-http"

[tool.poetry.dependencies]
python = "<3.13, >=3.12"
hikari = {git = "https://github.com/hikari-py/hikari.git", rev = "refs/pull/1803/merge"}
hikari-yuyo = { version = "1.19.1", extras = ["asgi"] }
motor = "3.3.2"
hiredis = "2.3.2"
redis = "5.0.1"
requests = "2.31.0"
dnspython = "2.5.0"
psutil = "5.9.0"
python-dateutil = "2.8.2"
blacksheep = "2.0.6"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
python-dotenv = "^1.0.1"
asgi_prometheus = "1.1.2"
bloxlink-lib = {git = "https://github.com/bloxlink/bloxlink-lib"}
pydantic = "^2.6.0"

[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
isort = "^5.12.0"
pylint = "^3.0.1"

[tool.black]
line-length = 110

[tool.isort]
profile = "black"
line_length = 110

[tool.pylint]
max-line-length = 110
disable = ["C0413", "C0103", "C0114", "C0301", "W1203"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
